1. retirado 'action'; 2. modificações no template de test
