+++ pytrobot/cli/main.py
@@ -5,14 +5,15 @@ from invoke.collection import Collection
 from invoke.program import Program
 
 # Importe suas tarefas aqui
-from pytrobot.tasks import build, new, tool, action, state
+from pytrobot.tasks import build, new, tool, state, testState
 
 namespace = Collection()
 namespace.add_task(build, name="build") #type:ignore
 namespace.add_task(new, name="new") #type:ignore
 namespace.add_task(tool, name="tool") #type:ignore
-namespace.add_task(action, name="action") #type:ignore
 namespace.add_task(state, name="state") #type:ignore
+namespace.add_task(testState, name="testState") #type:ignore
+
 
 # Adicione mais tarefas conforme necessário
