--- pytrobot/scaffold/src/actions/__init__.py
@@ -1,1 +1,0 @@
-from pytrobot.scaffold.src.actions.sample_action import SampleAction
